fixing docs
Former-commit-id: fb1972c5f6dcf2e8570b2d89fcdc5c86ca42b28a

+++ source/class_environment.h
@@ -19,28 +19,11 @@ private:
 public:
     std::string single_picture = "=";
     std::string full_picture;
-    size_t size;         //Size is the number of particles that have been contracted into this environment
+    int size;         //Size is the number of particles that have been contracted into this environment
 
     Tensor3 block;
     class_environment_L();
-//    class_environment_L& operator=(class_environment_L other){
-//        std::swap(*this, other); // (2)
-//
-//        return *this;
-//    }
-    // move constructor
-//    class_environment_L(class_environment_L&& other): class_environment_L() // initialize via default constructor, C++11 only
-//    {
-//        std::swap(*this, other);
-//    }
-
-
-
     void enlarge(const class_MPS &MPS, const Tensor4 &W);
-
-//    void enlarge_left(const Gamma_AB &G, const Lambda_AB &L, const Tensor4 &W);
-//    void enlarge_right(const Gamma_AB &G, const Lambda_AB &L, const Tensor4 &W);
-
 };
 
 
@@ -50,29 +33,12 @@ private:
 public:
     std::string single_picture = "-";
     std::string full_picture;
-    size_t size;         //Size is the number of particles that have been contracted into this environment
+    int size;         //Size is the number of particles that have been contracted into this environment
 
     Tensor3 block;
     class_environment_R();
 
-
-//    class_environment_R& operator=(class_environment_R other){
-//        std::swap(*this, other); // (2)
-//
-//        return *this;
-//    }
-    // move constructor
-//    class_environment_R(class_environment_R&& other): class_environment_R() // initialize via default constructor, C++11 only
-//    {
-//        std::swap(*this, other);
-//    }
-
-
     void enlarge(const class_MPS &MPS,const Tensor4 &W);
-
-//    void enlarge_left(const Gamma_AB &G, const Lambda_AB &L, const Tensor4 &W);
-//    void enlarge_right(const Gamma_AB &G, const Lambda_AB &L, const Tensor4 &W);
-
 };
